+++ bunker/src/app/context/client-provider.js
@@ -1,0 +1,11 @@
+'use client'
+
+import { SessionProvider } from "next-auth/react"
+
+export default function Provider ({children, session}){
+  return (
+    <SessionProvider session={session}>
+        {children}
+    </SessionProvider>
+  )
+}
